refractor done, need to fix norm and put header
