+++ Makefile
@@ -41,7 +41,10 @@ SRCS_FILES		=	error.c \
 					expander2.c	\
 					quote_removal.c \
 					history.c \
+					history_utils.c \
+					history_file.c \
 					string_utils.c
+
 HEADERS			=	$(INC_DIR)/minishell.h \
 					$(SRCS_DIR)/log.h \
 					$(SRCS_DIR)/parser.h
