+++ srcs/history.c
@@ -10,12 +10,7 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
 #include "minishell.h"
-#include <fcntl.h>
-static int	is_empty_line(char *line);
-static char	*read_line_from_fd(int fd);
-static void	load_history_from_file(t_history *hist);
 
 t_history	*init_history(void)
 {
@@ -30,6 +25,16 @@ t_history	*init_history(void)
 	load_history_from_file(hist);
 	return (hist);
 }
+
+static void	write_command_to_file(int fd, char *command)
+{
+	if (command)
+	{
+		write(fd, command, ft_strlen(command));
+		write(fd, "\n", 1);
+	}
+}
+
 void	save_to_history_file(t_history *hist)
 {
 	int	fd;
@@ -46,162 +51,9 @@ void	save_to_history_file(t_history *hist)
 	i = 0;
 	while (i < hist->count)
 	{
-		if (hist->commands[i])
-		{
-			write(fd, hist->commands[i], ft_strlen(hist->commands[i]));
-			write(fd, "\n", 1);
-		}
+		write_command_to_file(fd, hist->commands[i]);
 		i++;
 	}
 	close (fd);
 }
 
-void	add_to_history_struct(char *line, t_history *hist)
-{
-	char	*tmp;
-
-	if (!line || !hist || is_empty_line(line))
-		return ;
-	if (hist->count > 0 && hist->commands[hist->count - 1])
-	{
-		if (!ft_strncmp(hist->commands[hist->count - 1], line, ft_strlen(line)))
-			return ;
-	}
-	add_history(line);
-	if (hist->count >= hist->capacity)
-	{
-		if(hist->capacity == 0)
-		{
-			hist->capacity = INITIAL_CAPACITY;
-		}
-		else
-			hist->capacity = hist->capacity * 2;
-		tmp = ft_realloc(hist->commands, hist->capacity * sizeof(char *));
-		if (!tmp)
-			return ;
-		hist->commands = (char **)tmp;
-	}
-	hist->commands[hist->count] = ft_strdup(line);
-	if (hist->commands[hist->count])
-		hist->count++;
-}
-
-void	free_history_struct(t_history *hist)
-{
-	int	i;
-
-	if (!hist)
-		return ;
-	if (hist->commands)
-	{
-		i = 0;
-		while (i < hist->count)
-		{
-			if (hist->commands[i])
-				free(hist->commands[i]);
-			i++;
-		}
-		free(hist->commands);
-	}
-}
-
-static void	load_history_from_file(t_history *hist)
-{
-	int		fd;
-	char	*line;
-
-	if (!hist)
-		return ;
-	if (access(HISTORY_FILE, F_OK) != 0)
-		return ;
-	fd = open(HISTORY_FILE, O_RDONLY);
-	if (fd == -1)
-	{
-		perror("Error opening history file");
-		return ;
-	}
-	while ((line = read_line_from_fd(fd)) != NULL)
-	{
-		if (!is_empty_line(line))
-		{
-			add_history(line);
-			if (hist->count >= hist->capacity)
-			{
-				if (hist->capacity == 0)
-				{
-					hist->capacity = INITIAL_CAPACITY;
-				}
-				else
-					hist->capacity = hist->capacity * 2;
-				hist->commands = ft_realloc(hist->commands, hist->capacity * sizeof(char *));
-				if (!hist->commands)
-				{
-					free(line);
-					close(fd);
-					return ;
-				}
-			}
-			hist->commands[hist->count] = ft_strdup(line);
-			if (hist->commands[hist->count])
-				hist->count++;
-		}
-		free(line);
-	}
-	close(fd);
-}
-
-static char	*read_line_from_fd(int fd)
-{
-	char	*line;
-	char	buffer;
-	int		len;
-	int		capacity;
-	char	*tmp;
-
-	len = 0;
-	capacity = 64;
-	line = malloc(capacity + 1);
-	if (!line)
-		return (NULL);
-	while (read(fd, &buffer, 1) > 0)
-	{
-		if (buffer == '\n')
-			break;
-		if (len >= capacity)
-		{
-			capacity = capacity * 2;
-			tmp = ft_realloc(line, capacity + 1);
-			if (!tmp)
-			{
-				free(line);
-				return (NULL);
-			}
-			line = tmp;
-		}
-		line[len] = buffer;
-		len++;
-	}
-	line[len] = '\0';
-	if (len == 0)
-	{
-		free(line);
-		return (NULL);
-	}
-	return (line);
-}
-
-static int	is_empty_line(char *line)
-{
-	int	i;
-
-	if(!line)
-		return (1);
-	i = 0;
-	while (line[i])
-	{
-		if (line[i] != ' ' && line[i] != '\t' && line[i] != '\n')
-			return (0);
-		i++;
-	}
-	return (1);
-}
